compel: Save thread registers before executing parasite code
For dumping threads we execute parasite code before we collect the
floating point registers with get_task_regs().

The following describes the code path were this is done:

 dump_task_threads()
     for (i = 0; i < item->nr_threads; i++)
         dump_task_thread()
           parasite_dump_thread_seized()
             compel_prepare_thread()
                prepare_thread()

                   ### Get general purpose registers ###
                   ptrace_get_regs(ctx->regs)

             ### Parasite code is executed in thread ###
             compel_run_in_thread(tctl, PARASITE_CMD_DUMP_THREAD)

             compel_get_thread_regs()

                 ### Get FP and VX registers ###
                 get_task_regs()

On s390 gcc uses floating point registers to cache general purpose
registers. Therefore we use the -msoft-float option to ensure that
floating point registers are not clobbered in compel_run_in_thread().

There might be a better option:

First save the thread registers and then call the parasite code. This
allows us to remove the -msoft-float on s390.

Signed-off-by: Michael Holzheu <holzheu@linux.vnet.ibm.com>