Fix warning reported by openSUSE's OBS (Open Build Service)
Hi,

Please find the following patch attached, which removes the shebang

        -#!/bin/env python2

from a single file: ( ./lib/py/images/images.py )

If I am understanding things correctly, images.py can never be
executed on it's own (i.e., via "./images.py"), so having the shebang
is superfluous.

It's a relatively trivial thing but would be nice to have this.

Here, you can see the build result for the suse rpm is clean:

https://build.opensuse.org/request/show/541010

The failures on travis appear to be unrelated, but I'd appreciate a
pair of more finely tuned eyes there:

https://travis-ci.org/checkpoint-restore/criu/builds/299831055

Thanks for reading.

Regards,

Avindra.

>From aee84030f4623c468faaa90b8fc0820e39f86bda Mon Sep 17 00:00:00 2001
From: Avindra Goolcharan <aavindraa@gmail.com>
Date: Sat, 4 Nov 2017 18:46:03 -0400
Subject: [PATCH] images.py: remove shebang

This file is not executable directly, so it should not have the shebang.