Track which configuration options have been changed
For the upcoming RPC configuration file support, the RPC code path needs
to know which values have been changed from the default.

The reason for this is, that we do not want that values specified via
RPC are overwriting user specified settings in the configuration file.

For bool options is is not possible to tell if the value is the default
value as a result of the configuration file or as a result of just being
the default.

For default values, which are not modified by configuration files we
want to use the RPC specified value. If a user changes a value to the
default in the configuration file we do not want to have it changed by
the RPC code path.

This means that values from the configuration file have a higher
priority than values configured via RPC.

The reason for this is, that we want to enable the user to change CRIU's
behavior when embedded in a container runtime and used via RPC (e.g.
runc).

If an option is specified in the configuration file the values for that
option are set to 1/true in the corresponding opt_changed_by_config.

Now the RPC code path can check if an option has been changed via
configuration file and can then ignore the value specified via RPC.

Signed-off-by: Adrian Reber <areber@redhat.com>