RPC: Use configuration file also from RPC
This enables the usage of configuration file values in the RPC code
path. The main change is to verify if the user has actually changed a
configuration option in the configuration file and if it has been
changed the value configured via RPC is ignored.

As it is not possible to detect if a configuration option has the
default value because it has not been changed or set by the user or if
the user actually set it to the default value a copy of the options
structure (opt_changed_by_config) is used to track if a configuration
option has been changed by the user. This structure does not contain
that actual value of the configuration option it only tracks if an
option has been set. The value is still in the original options
structure (opts).

With this it is now possible to override CRIU's behavior if started via
RPC. If the user sets a value in one of the configuration files the
value coming from RPC is ignored. This can lead to undesired behavior if
the configuration file sets something which breaks what the RPC client
wants to do. On the other hand it gives the user flexibility to change
CRIU's behavior without potentially having to change the code of the RPC
client.

Signed-off-by: Adrian Reber <areber@redhat.com>