stats/pid-reuse: put stats to image directory instead of cwd
Statistics for a special dump/pre-dump/restore action should be
collected in the images directory of these action. We need these for
pid-reuse detection as we need to read the stats of parent predump.

Now then work-dir option is set statistics image is put to the work-dir
and on each iteration these image will replace previos image.

For backward compatibility with older versions of p.haul create a
symlink in cwd so that p.haul still has access to the latest stats
image. Yet I plan to make p.haul work with stats through images dir.

https://jira.sw.ru/browse/PSBM-82864
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>