img-cache: Daemonize on start
When the --daemon option is passed to image-cache, it will wait until a
connection with image-proxy is established, and then will continue its
execution in the background (daemonize).

This commit will daemonize image-cache immediately after start. This
enables the following two commands to be executed sequentially (in a
single terminal window):

$ criu image-cache -d --port 1234 -o img-cache.log
$ criu image-proxy -d --port 1234 --address 0.0.0.0 -o img-proxy.log

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>