config: check for CRIU_CONFIG_FILE environment variable
To be able to influence, especially via RPC, that another configuration
file should be used, this introduces the environment variable
CRIU_CONFIG_FILE.

If set, the file it points to will be used instead of the default
configuration files and also instead of the configuration file specified
via the command-line.

CRIU now first checks for CRIU_CONFIG_FILE and uses that, if that is not
set it looks for '--config FILEPATH' and uses that, if that is not set
it uses the default configuration files.

The idea behind this option is that now it is possible for an RPC client
to set CRIU_CONFIG_FILE and also set the RPC option
opts.prefer_config_file to tell CRIU to prefer the options from a
non-default configuration file over the options set via RPC.

Signed-off-by: Adrian Reber <areber@redhat.com>