Adding checkpoint/restore support to podman question
On Wed, Jul 18, 2018 at 03:25:09PM +0200, Adrian Reber wrote:
> > On Fri, May 04, 2018 at 02:01:08PM -0700, Andrei Vagin wrote:
> > > On Thu, May 03, 2018 at 07:48:55PM +0200, Adrian Reber wrote:
> > > > If you could implement the CRIU part that would be great. I have not
> > > > started yet but I would then do the necessary changes in runc (and
> > > > higher).
> > >
> > > Here is a criu part:
> > > https://github.com/avagin/criu/tree/netns_ext
> > >
> > > Let me know if you will have any questions or comments.

As your patch does not apply anymore (I guess due to the removal of
nested namespaces) I tried to port your patch to the current criu-dev
branch. And it almost works. I am storing the ext_key now in
'netns_entry', but the problem is that that image file is read to late.

So at the point the network namespace is restored the necessary
information is not yet read from the image file and the network
namespace is never restored as described via inherit_fd.

Please have a look at my attached patch and maybe you have an idea how
to correctly implement it. Maybe you could just let me know which image
file would be best suited for the ext_key of the network namespace and
how to make sure it is available at restore time. Thanks.

		Adrian
Acked-by: Adrian Reber <areber@redhat.com>