action-scripts: add defines for inherit fd over RPC
To be able to use inherit-fd over RPC, the RPC client will only submit
the key to CRIU via RPC. The actual FD which relates to this key has to
be transmitted directly and not via a protobuf message.

The idea is to let CRIU send a 'notify' message to the RPC client and
ask for the FD to be transmitted over the unix domain socket between
CRIU and the RPC client.

 * RPC client requests restore with inherit-fd using 'key'
 * CRIU asks RPC client for the actual FD via a notify protobuf message
 * RPC client transmits the FD to CRIU

As the protobuf notify mechanism naming is based on the action scripts
the naming of the protobuf fields is not perfect. The field with the
notify functionality is called 'script', but we have to live with it
now.

Signed-off-by: Adrian Reber <areber@redhat.com>