cr-service: use inherit-fd keys to ask for FD
Using inherit-fd in RPC mode uses the notify mechanism to ask the RPC
client for the real FD.

The basic workflow of inherit-fd in RPC mode is:

 * RPC clients fills in inherit-fd field:
   * inheritFd.Key = proto.String(path.Base(nsPath))
   * inheritFd.Fd is irrelevant as it will be transmitted later
 * CRIU receives RPC messages and calls inherit_fd_add_rpc()
   * this transmits the inheritFd.Key back to the RPC client
   * RPC client uses unix domain socket to send the real FD
     to CRIU

Unfortunately this is more complicated as runc already uses
inheritFd.Fd to send the file descriptors for 0, 1, 2. As those are
always stdin, stdout, stderr is just worked even if the transmitted
values via RPC were not actual FDs but only numbers.

To handle this, CRIU still accepts inherit-fds directly as FDs if it is
0, 1, 2. This is a compatibility hack to avoid breaking runc.

Signed-off-by: Adrian Reber <areber@redhat.com>