files: Simplify get_mem_dev_ops
The former 'case' statement was used to allow the code to be easily
extended in the future (e.g. if there are more minor numbers to
be added we would not have to modify much code but simply add a new
'case'). However, we don't expect such extension in the near future.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>
Acked-by: Cyrill Gorcunov <gorcunov@gmail.com>